Add support for client action menu in timers (tvh4.3+).
This allows the user to mark a recording as "previously recorded".

The existing "disable timer" is useful for manually scheduled
recordings. However timers that are created by autorec rules can often
match recordings the user does not wish to record.

For example "record all Tom Cruise" would enable a timer for "Mission
Impossible" every time the programme is aired. Although the user can
disable the timer for a specific airdate, the next time the programme
is aired a new timer will be scheduled to record the programme.

By contrast, "previously recorded" flags the scheduled item as already
recorded so that the airing and any future airings are marked as
"duplicate, do not record", according to the user's retention/dup
policy.